decoupled account compomnents
