+++ frontend/app/features/users/types.ts
@@ -1,2 +1,6 @@
+export type SideBarSection = 'User Information' | 'Event Bookings' | 'Personal Calendar' | 'Chat Room';
 
-export type SideBarSection = 'Event Bookings' | 'Personal Calendar' | 'Chat Room';
+export interface ProfileMainContentProps {
+  section: SideBarSection;
+}
+
